+++ src/components/detail/footer/Footer.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import styled from 'styled-components';
-import DetailCard from '../footer/DetailCard';
+import DetailCard from '../../common/Card';
 import PlusBtn from '../../../assets/icons/PlusBtn.svg';
 import FooterMenu from './FooterMenu';
+import imgUrl from '../../../assets/icons/example.jpeg';
 
 function Footer() {
-  /*임의로 20개의 데이터를 넣음*/
-  const cardData = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19];
+  const cardData = {
+    view: 'detailDesktop',
+    /*임의의 데이터 넣음*/
+    imgUrl: imgUrl,
+    category: '구단특집',
+    title: '[IUFC TV] 3 - 1 승리! 원정 첫 승리를 팬분들께 바칩니다!  승리의 무인 퇴근캠',
+    createdAt: '인천유나이티드 IUFC TV',
+    data: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19],
+  };
 
   return (
     <>
@@ -16,7 +24,9 @@ function Footer() {
           <div className="introduce__detail">주제별로 분류된 다양한 영상 모음</div>
         </div>
         {/* 4*5 grid 틀에 map 돌려서 DetailCard 넣기*/}
-        <CardList>{cardData && cardData.map(() => <DetailCard />)}</CardList>
+        <CardList>
+          {cardData.data && cardData.data.map(() => <DetailCard cardData={cardData} />)}
+        </CardList>
         <div className="more">
           <img src={PlusBtn} className="more--button" alt="" />
           <div className="more--text">새로운 글 더보기</div>
