+++ src/components/main/footer/Footer.js
@@ -1,26 +1,25 @@
 import React from 'react';
-import MainCard from '../footer/MainCard';
+import MainCard from '../../common/Card';
 import SettingIcon from '../../../assets/icons/Setting.svg';
 import LeftButton from '../../../assets/icons/LeftButton.svg';
 import RightButton from '../../../assets/icons/RightButton.svg';
 import styled from 'styled-components';
 import FooterMenu from './FooterMenu';
 import { useHistory } from 'react-router-dom';
+import imgUrl from '../../../assets/icons/example2.jpeg';
 
 function Footer() {
   const history = useHistory();
   // 임의의 데이터 넣기 -> 서버 연결하면 데이터 받아서 넣기
+  //서버로부터 받을 데이터 -> imgUrl, category,title, createdAt
+  //우리가 모듈화할 데이터 mainDesktop , mainTablet , detailDesktop , detailTab
   const cardData = {
-    allWidth: '37.8rem',
-    cardHeight: '48.8rem',
-    imageHeight: '37.8rem',
-    kindFontSize: '1.6rem',
-    titleFontSize: '1.6rem',
-    categoryFontSize: '1.5rem',
+    view: 'mainDesktop',
     /*임의의 데이터 넣음*/
-    kind: '구단특집',
+    imgUrl: imgUrl,
+    category: '구단특집',
     title: '[IUFC TV] 3 - 1 승리! 원정 첫 승리를 팬분들께 바칩니다!  승리의 무인 퇴근캠',
-    category: '인천유나이티드 IUFC TV',
+    createdAt: '인천유나이티드 IUFC TV',
     data: [0, 1, 2],
   };
   const clickHandler = () => {
